module_c_integration/api/requirements.txt
fastapi>=0.116.1
uvicorn[standard]>=0.27.1
pydantic>=2.10.3
pydantic-settings>=2.1.0
motor>=3.3.2           # Async MongoDB driver
pymongo>=4.6.1
web3>=6.14.0           # Ethereum interactions
python-jose[cryptography]>=3.3.0  # JWT
passlib[bcrypt]>=1.7.4            # Password hashing
python-multipart>=0.0.9           # File uploads
httpx>=0.28.1          # Async HTTP client (calls Module A Vision API)
ipfshttpclient==0.8.0a2
aiofiles>=23.2.1
python-dotenv>=1.0.1
PyYAML>=6.0.2
websockets>=12.0
pytest>=7.4.4
pytest-asyncio>=0.23.4
pytest-cov>=4.1.0
flake8>=7.0.0
black>=24.1.1
